INKLB-004 Move websocket concepts to separate header
